+++ lib/vagrant/environment.rb
@@ -139,6 +139,7 @@ module Vagrant
     #
     # @return [Array<VM>]
     def vms_ordered
+      return @vms.values if !multivm?
       @vms_enum ||= config.global.vm.defined_vm_keys.map { |name| @vms[name] }
     end
 
@@ -172,7 +173,7 @@ module Vagrant
     #     env.cli("package", "--vagrantfile", "Vagrantfile")
     #
     def cli(*args)
-      CLI.start(args.flatten, :env => self)
+      CLI.new(args.flatten, self).execute
     end
 
     # Returns the host object associated with this environment.
@@ -359,7 +360,13 @@ module Vagrant
       config_loader = Config::Loader.new
       config_loader.load_order = [:default, :box, :home, :root, :vm]
 
-      inner_load = lambda do |subvm=nil, box=nil|
+      inner_load = lambda do |*args|
+        # This is for Ruby 1.8.7 compatibility. Ruby 1.8.7 doesn't allow
+        # default arguments for lambdas, so we get around by doing a *args
+        # and setting the args here.
+        subvm = args[0]
+        box   = args[1]
+
         # Default Vagrantfile first. This is the Vagrantfile that ships
         # with Vagrant.
         config_loader.set(:default, File.expand_path("config/default.rb", Vagrant.source_root))