+++ lib/vagrant/provisioners/chef_client.rb
@@ -1,12 +1,12 @@
 require 'pathname'
 
+require 'vagrant/provisioners/chef'
+
 module Vagrant
   module Provisioners
     # This class implements provisioning via chef-client, allowing provisioning
     # with a chef server.
     class ChefClient < Chef
-      register :chef_client
-
       class Config < Chef::Config
         attr_accessor :chef_server_url
         attr_accessor :validation_key_path
@@ -29,13 +29,17 @@ module Vagrant
           @encrypted_data_bag_secret = "/tmp/encrypted_data_bag_secret"
         end
 
-        def validate(errors)
+        def validate(env, errors)
           super
 
           errors.add(I18n.t("vagrant.config.chef.server_url_empty")) if !chef_server_url || chef_server_url.strip == ""
           errors.add(I18n.t("vagrant.config.chef.validation_key_path")) if !validation_key_path
           errors.add(I18n.t("vagrant.config.chef.run_list_empty")) if @run_list && @run_list.empty?
         end
+      end
+
+      def self.config_class
+        Config
       end
 
       def prepare
@@ -56,22 +60,22 @@ module Vagrant
       end
 
       def create_client_key_folder
-        env.ui.info I18n.t("vagrant.provisioners.chef.client_key_folder")
+        env[:ui].info I18n.t("vagrant.provisioners.chef.client_key_folder")
         path = Pathname.new(config.client_key_path)
 
-        vm.ssh.execute do |ssh|
+        env[:vm].ssh.execute do |ssh|
           ssh.sudo!("mkdir -p #{path.dirname}")
         end
       end
 
       def upload_validation_key
-        env.ui.info I18n.t("vagrant.provisioners.chef.upload_validation_key")
-        vm.ssh.upload!(validation_key_path, guest_validation_key_path)
+        env[:ui].info I18n.t("vagrant.provisioners.chef.upload_validation_key")
+        env[:vm].ssh.upload!(validation_key_path, guest_validation_key_path)
       end
 
       def upload_encrypted_data_bag_secret
-        env.ui.info I18n.t("vagrant.provisioners.chef.upload_encrypted_data_bag_secret_key")
-        vm.ssh.upload!(encrypted_data_bag_secret_key_path, config.encrypted_data_bag_secret)
+        env[:ui].info I18n.t("vagrant.provisioners.chef.upload_encrypted_data_bag_secret_key")
+        env[:vm].ssh.upload!(encrypted_data_bag_secret_key_path, config.encrypted_data_bag_secret)
       end
 
       def setup_server_config
@@ -92,24 +96,29 @@ module Vagrant
         command_env = config.binary_env ? "#{config.binary_env} " : ""
         command = "#{command_env}#{chef_binary_path("chef-client")} -c #{config.provisioning_path}/client.rb -j #{config.provisioning_path}/dna.json"
 
-        env.ui.info I18n.t("vagrant.provisioners.chef.running_client")
-        vm.ssh.execute do |ssh|
+        env[:ui].info I18n.t("vagrant.provisioners.chef.running_client")
+        env[:vm].ssh.execute do |ssh|
           ssh.sudo!(command) do |channel, type, data|
             if type == :exit_status
               ssh.check_exit_status(data, command)
             else
-              env.ui.info("#{data}: #{type}")
+              # Output the data with the proper color based on the stream.
+              color = type == :stdout ? :green : :red
+
+              # Note: Be sure to chomp the data to avoid the newlines that the
+              # Chef outputs.
+              env[:ui].info(data.chomp, :color => color, :prefix => false)
             end
           end
         end
       end
 
       def validation_key_path
-        File.expand_path(config.validation_key_path, env.root_path)
+        File.expand_path(config.validation_key_path, env[:root_path])
       end
 
       def encrypted_data_bag_secret_key_path
-        File.expand_path(config.encrypted_data_bag_secret_key_path, env.root_path)
+        File.expand_path(config.encrypted_data_bag_secret_key_path, env[:root_path])
       end
 
       def guest_validation_key_path